# docs/user/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.responseLatestAritcles:
    properties:
      articles:
        items:
          $ref: '#/definitions/model.Article'
        type: array
      pager:
        $ref: '#/definitions/pager.Pager'
    type: object
...
  model.Article:
    properties:
      category:
        $ref: '#/definitions/model.Category'
      categoryId:
        type: integer
      content:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      status:
        type: integer
      tags:
        items:
          $ref: '#/definitions/model.Tag'
        type: array
      thumbnailUrl:
        type: string
      title:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/model.User'
      userId:
        type: integer
    type: object
  model.Category:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      slug:
        type: string
    type: object
  model.Tag:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      slug:
        type: string
    type: object
  model.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
    type: object
  pager.Pager:
    properties:
      currentPage:
        type: integer
      lastPage:
        type: integer
    type: object
  request.JSONResponce:
    properties:
      data: {}
      status:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
...
  /api/new/{page}:
    get:
      consumes:
      - application/json
      description: get the latest articles by page
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/request.JSONResponce'
            - properties:
                data:
                  $ref: '#/definitions/handler.responseLatestAritcles'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/request.JSONResponce'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/request.JSONResponce'
            - properties:
                data:
                  type: string
              type: object
      summary: latest_articles handlers
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"